[modified] AuctionItem to use idlc and updated Bazaar code to use it [added] Basic Bazaar auction code from starvin ticket #68
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@221 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/bazaar/BazaarPlanetManagerImplementation.h
@@ -67,7 +67,7 @@ public:
 	
 	void setPlanet(int planet);
 	void addBazaarItem(AuctionItem* auctionItem);
-	void removeBazaarItem(long long objectid);
+	void removeBazaarItem(uint64 objectid);
 	void addBazaar(BazaarTerminalDetails* terminal);
 };
